pyproject.toml
[project]
name = "build-your-first-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
]
